+++ kernel/src/lib.rs
@@ -15,6 +15,7 @@ pub mod initramfs;
 pub mod memory;
 pub mod process_image;
 pub mod user;
+pub mod user_copy;
 pub mod user_stack;
 
 pub mod arch;